fabric/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
blinker==1.4
certifi==2020.6.20
chardet==3.0.4
charset-normalizer==3.1.0
click==7.1.2
colorama==0.4.6
distlib==0.3.6
Django==4.1.7
et-xmlfile==1.1.0
filelock==3.8.0
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
greenlet==2.0.2
idna==2.10
importlib-metadata==5.0.0
inflect==4.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
openpyxl==3.1.2
paytmchecksum==1.7.0
platformdirs==2.5.3
psycopg2==2.9.7
pycryptodome==3.12.0
pyfcm==1.5.1
PyJWT==1.7.1
pyodbc==4.0.30
python-dotenv==0.14.0
PyYAML==5.3.1
queries==2.1.1
redis==3.5.3
requests==2.24.0
six==1.15.0
sqlacodegen==2.3.0
SQLAlchemy==1.3.19
sqlparse==0.4.3
typing-extensions==4.7.1
tzdata==2022.7
ua-parser==0.10.0
urllib3==1.25.10
user-agents==2.2.0
virtualenv==20.16.6
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.10.0
